+++ src/Navigation/Page.php
@@ -29,7 +29,9 @@ class Page extends \KodiComponents\Navigation\Page implements PageInterface
         $this->setModel($modelClass);
 
         if ($this->hasModel()) {
-            $this->setIcon($this->getModelConfiguration()->getIcon());
+            if ($this->getModelConfiguration()->getIcon()) {
+                $this->setIcon($this->getModelConfiguration()->getIcon());
+            }
         }
     }